@@ -47,97 +47,171 @@
   },
   "homepage": "https://github.com/seek-oss/seek-style-guide#readme",
   "dependencies": {
-    "autoprefixer": "6.7.7",
-    "babel-core": "^6.10.4",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-add-react-displayname": "^0.0.4",
-    "babel-plugin-transform-class-properties": "^6.10.2",
-    "babel-plugin-transform-object-rest-spread": "^6.8.0",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.11.1",
-    "basekick": "3.0.0",
-    "chalk": "1.1.3",
+    "basekick": "^3.0.0",
     "classnames": "^2.2.5",
-    "css-loader": "0.28.0",
-    "enzyme": "^2.8.2",
-    "enzyme-to-json": "^1.5.1",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.9.0",
-    "less": "2.7.1",
-    "less-loader": "2.2.3",
-    "lodash": "4.17.4",
+    "lodash": "^4.17.4",
     "pad-left": "^2.1.0",
-    "postcss-loader": "0.9.1",
-    "prop-types": "^15.5.10",
-    "raw-loader": "0.5.1",
-    "react-autosuggest": "9.3.1",
-    "react-isolated-scroll": "^0.1.0",
-    "react-scrolllock": "1.0.8",
-    "style-loader": "0.13.1",
-    "svgo": "0.7.1",
-    "svgo-loader": "1.1.0",
-    "uglify-loader": "1.3.0",
-    "url-loader": "0.5.7",
-    "webpack-node-externals": "1.2.0"
+    "prop-types": "^15.6.0",
+    "react-autosuggest": "^9.3.2",
+    "react-isolated-scroll": "^0.1.1",
+    "react-scrolllock": "^2.0.1"
   },
   "peerDependencies": {
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
+    "react": "^15.0.0 || ^16.0.0",
+    "react-dom": "^15.0.0 || ^16.0.0",
     "react-helmet": "^3.0.0 || ^4.0.0 || ^5.0.0"
   },
   "devDependencies": {
-    "babel-eslint": "^6.0.4",
-    "babel-jest": "^18.0.0",
-    "chai": "^3.5.0",
-    "commitizen": "^2.9.5",
-    "cz-conventional-changelog": "^2.0.0",
-    "ejs": "^2.4.2",
-    "es6-promisify": "^5.0.0",
-    "eslint": "^3.17.1",
-    "eslint-plugin-babel": "^4.0.1",
-    "eslint-plugin-jest": "^1.0.2",
-    "eslint-plugin-react": "^6.0.0-alpha.1",
-    "extract-zip": "^1.6.5",
-    "gh-pages": "^0.11.0",
-    "glob": "^7.1.1",
+    "@commitlint/cli": "^5.2.8",
+    "@commitlint/config-conventional": "^5.2.3",
+    "autoprefixer": "^7.1.5",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.1",
+    "babel-jest": "^21.2.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-add-react-displayname": "^0.0.4",
+    "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-react": "^6.24.1",
+    "browserslist-config-seek": "^1.0.0",
+    "canvg-fixed": "^1.0.0",
+    "chai": "^4.1.2",
+    "chalk": "^2.1.0",
+    "commitizen": "^2.9.6",
+    "css-loader": "^0.28.7",
+    "cz-conventional-changelog": "^2.1.0",
+    "dedent": "^0.7.0",
+    "ejs": "^2.5.7",
+    "enzyme": "^3.1.0",
+    "enzyme-adapter-react-16": "^1.0.1",
+    "enzyme-to-json": "^3.1.2",
+    "eslint": "^4.8.0",
+    "eslint-plugin-babel": "^4.1.2",
+    "eslint-plugin-jest": "^21.2.0",
+    "eslint-plugin-react": "^7.4.0",
+    "extract-text-webpack-plugin": "^3.0.1",
+    "gh-pages": "^1.0.0",
     "hex-rgb": "^1.0.0",
-    "html-webpack-plugin": "^2.28.0",
-    "husky": "^0.12.0",
+    "html-minifier": "^3.5.6",
+    "html-sketchapp-cli": "0.0.2",
+    "html-webpack-plugin": "^2.30.1",
+    "husky": "^0.14.3",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^19.0.0",
-    "json-loader": "^0.5.4",
-    "jsx-to-string": "^0.2.13",
-    "less-vars-to-js": "^1.1.1",
-    "opn": "^4.0.1",
+    "import-glob": "^1.5.0",
+    "jest": "^22.0.0",
+    "json-loader": "^0.5.7",
+    "jsx-to-string": "^1.1.0",
+    "less": "^2.7.2",
+    "less-loader": "^4.0.5",
+    "less-vars-to-js": "^1.2.0",
+    "module-alias": "^2.0.1",
+    "opn": "^5.1.0",
     "opn-cli": "^3.1.0",
+    "postcss-loader": "^2.0.6",
     "postcss-prefix-selector": "^1.6.0",
-    "react": "^15.3.1",
-    "react-addons-pure-render-mixin": "^15.3.1",
-    "react-baseline": "^1.0.1",
-    "react-copy-to-clipboard": "4.2.1",
-    "react-dom": "^15.3.1",
-    "react-element-to-jsx-string": "^12.0.0",
-    "react-helmet": "^3.2.2",
-    "react-hot-loader": "^3.0.0-beta.2",
-    "react-router": "^2.5.2",
-    "react-shallow-testutils": "^2.0.0",
-    "react-test-renderer": "^15.4.2",
-    "rewrite-files": "^1.0.1",
-    "rimraf": "^2.6.1",
-    "semantic-release": "^6.3.2",
-    "sinon": "^1.17.4",
-    "sinon-chai": "^2.8.0",
+    "raf": "^3.4.0",
+    "raw-loader": "^0.5.1",
+    "react": "^16.0.0",
+    "react-copy-to-clipboard": "^5.0.1",
+    "react-dom": "^16.0.0",
+    "react-element-to-jsx-string": "^13.0.0",
+    "react-helmet": "^5.2.0",
+    "react-hot-loader": "^3.0.0-beta.7",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
+    "react-shallow-testutils": "^3.0.0",
+    "react-test-renderer": "^16.0.0",
+    "seek-style-guide-webpack": "^2.1.1",
+    "semantic-release": "^8.1.2",
+    "serialize-javascript": "^1.4.0",
+    "sinon": "^4.0.1",
+    "sinon-chai": "^2.14.0",
     "static-site-generator-webpack-plugin": "^3.4.1",
-    "string-replace-loader": "^1.2.0",
-    "stylelint": "8.0.0",
-    "stylelint-config-css-modules": "^1.0.0",
-    "stylelint-config-standard": "^16.0.0",
-    "superstatic": "^4.0.3",
-    "surge": "^0.18.0",
-    "validate-commit-msg": "^2.8.2",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-fail-plugin": "^1.0.6",
-    "zip-folder": "^1.0.0"
+    "string-replace-loader": "^1.3.0",
+    "style-loader": "^0.19.0",
+    "stylelint": "^8.4.0",
+    "stylelint-config-css-modules": "^1.1.0",
+    "stylelint-config-standard": "^18.0.0",
+    "surge": "^0.19.0",
+    "svgo": "^1.0.0",
+    "svgo-loader": "^2.0.0",
+    "webpack": "^3.6.0",
+    "webpack-dev-server": "^2.9.1"
+  },
+  "greenkeeper": {
+    "ignore": [
+      "@commitlint/cli",
+      "@commitlint/config-conventional",
+      "autoprefixer",
+      "babel-core",
+      "babel-eslint",
+      "babel-jest",
+      "babel-loader",
+      "babel-plugin-add-react-displayname",
+      "babel-plugin-transform-class-properties",
+      "babel-plugin-transform-object-rest-spread",
+      "babel-preset-env",
+      "babel-preset-react",
+      "canvg-fixed",
+      "chai",
+      "chalk",
+      "commitizen",
+      "css-loader",
+      "cz-conventional-changelog",
+      "dedent",
+      "ejs",
+      "enzyme",
+      "enzyme-adapter-react-16",
+      "enzyme-to-json",
+      "eslint",
+      "eslint-plugin-babel",
+      "eslint-plugin-jest",
+      "eslint-plugin-react",
+      "extract-text-webpack-plugin",
+      "gh-pages",
+      "hex-rgb",
+      "html-minifier",
+      "html-sketchapp-cli",
+      "html-webpack-plugin",
+      "husky",
+      "identity-obj-proxy",
+      "import-glob",
+      "jest",
+      "json-loader",
+      "jsx-to-string",
+      "less",
+      "less-loader",
+      "less-vars-to-js",
+      "module-alias",
+      "opn",
+      "opn-cli",
+      "postcss-loader",
+      "postcss-prefix-selector",
+      "raf",
+      "raw-loader",
+      "react-copy-to-clipboard",
+      "react-element-to-jsx-string",
+      "react-hot-loader",
+      "react-router",
+      "react-router-dom",
+      "react-shallow-testutils",
+      "react-test-renderer",
+      "semantic-release",
+      "serialize-javascript",
+      "sinon",
+      "sinon-chai",
+      "static-site-generator-webpack-plugin",
+      "string-replace-loader",
+      "style-loader",
+      "stylelint",
+      "stylelint-config-css-modules",
+      "stylelint-config-standard",
+      "surge",
+      "svgo",
+      "svgo-loader",
+      "webpack",
+      "webpack-dev-server"
+    ]
   }
 }
