@@ -1,11 +1,3 @@
 module.exports = {
-  browsers: [
-    '> 1%',
-    'Last 2 versions',
-    'IE >= 10',
-    'Edge >= 13',
-    'Safari >= 6',
-    'iOS >= 6',
-    'Samsung >= 4'
-  ]
+  browsers: require('browserslist-config-seek')
 };
