@@ -1,20 +1,31 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-
+import { Route } from 'react-router';
 import { StyleGuideProvider } from 'seek-style-guide/react';
 import Header from './Header/Header';
 import Footer from './Footer/Footer';
+import Home from 'Home/Home';
+import PageLayout from 'PageLayout/PageLayout';
+import Typography from 'Typography/Typography';
+import Playground from 'Playground/Playground';
+import SketchExports from 'SketchExports/SketchExports';
+import Demo from 'Demo/Demo';
+import demoSpecExports from '../../../../react/*/*.demo.js';
+const demoSpecs = demoSpecExports.map(x => x.default);
 
-export default function App({ children }) {
-  return (
-    <StyleGuideProvider fullScreen={true} title="SEEK Style Guide">
-      <Header />
-      {children}
-      <Footer />
-    </StyleGuideProvider>
-  );
-}
+const demoRoutes = demoSpecs.map(demoSpec => {
+  const DemoRoute = () => <Demo spec={demoSpec} />;
+  return <Route key={demoSpec.title} path={demoSpec.route} component={DemoRoute} />;
+});
 
-App.propTypes = {
-  children: PropTypes.node
-};
+export default () => (
+  <StyleGuideProvider fullScreen={true} title="SEEK Style Guide">
+    <Header />
+    <Route path="/" exact component={Home} />
+    <Route path="/typography" component={Typography} />
+    <Route path="/page-layout" component={PageLayout} />
+    <Route path="/playground" component={Playground} />
+    <Route path="/sketch-exports" component={SketchExports} />
+    { demoRoutes }
+    <Footer />
+  </StyleGuideProvider>
+);
