@@ -9,7 +9,6 @@ import jsxToString from 'react-element-to-jsx-string';
 import CopyToClipboard from 'react-copy-to-clipboard';
 
 export default class Code extends Component {
-
   static propTypes = {
     jsx: PropTypes.element,
     less: PropTypes.string
@@ -51,10 +50,10 @@ export default class Code extends Component {
         filterProps: ['className', 'style'],
         useBooleanShorthandSyntax: false
       })
-      .replace(/\={true}/ig, '')
-      .replace(/svgClassName=".*?"/ig, 'svgClassName="..."')
-      .replace(/function noRefCheck\(\) \{\}/ig, '() => {...}')
-      .replace('<MockContent />', 'Lorem ipsum');
+        .replace(/\={true}/ig, '')
+        .replace(/svgClassName=".*?"/ig, 'svgClassName="..."')
+        .replace(/function noRefCheck\(\) \{\}/ig, '() => {...}')
+        .replace('<MockContent />', 'Lorem ipsum');
 
       const componentNames = uniq(
         (componentCode.match(/<([A-Z]\w*)(?=[\s>])/g) || [])
@@ -81,5 +80,4 @@ export default class Code extends Component {
       </CopyToClipboard>
     );
   }
-
 }
