@@ -4,7 +4,7 @@ import React from 'react';
 import lessToJs from 'less-vars-to-js';
 
 import Droplet from './Droplet/Droplet';
-import brand from '!!raw!seek-style-guide/theme/palette/brand.less';
+import brand from '!!raw-loader!seek-style-guide/theme/palette/brand.less';
 const brandValues = lessToJs(brand);
 
 const getSwatch = name => {
