@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 
 import lessToJs from 'less-vars-to-js';
 
-import grid from '!!raw!seek-style-guide/theme/layout/grid.less';
+import grid from '!!raw-loader!seek-style-guide/theme/layout/grid.less';
 const gridValues = lessToJs(grid);
 const rowHeight = parseInt(gridValues['@grid-row'], 10);
 
