@@ -14,7 +14,6 @@ import smoothScroll from '../private/smoothScroll';
 import smallDeviceOnly from '../private/smallDeviceOnly';
 
 export default class Autosuggest extends Component {
-
   static displayName = 'Autosuggest';
 
   static propTypes = {
@@ -91,22 +90,23 @@ export default class Autosuggest extends Component {
   }
 
   renderInputComponent = inputProps => {
-    const { labelProps } = this.props;
-
-    const enrichedLabelProps = {
-      ...labelProps,
-      className: classnames({
-        [styles.isLabelCoveredWithBackdrop]: this.props.showMobileBackdrop,
-        [labelProps.className]: labelProps.className
-      })
-    };
+    const { label, labelProps } = this.props;
 
     const textFieldProps = {
       ref: this.storeTextFieldReference,
       inputProps,
-      labelProps: enrichedLabelProps,
       ...omit(this.props, [ 'inputProps', 'labelProps', 'autosuggestProps' ])
     };
+
+    if (label) {
+      textFieldProps.labelProps = {
+        ...labelProps,
+        className: classnames({
+          [styles.isLabelCoveredWithBackdrop]: this.props.showMobileBackdrop,
+          [labelProps.className]: labelProps.className
+        })
+      };
+    }
 
     return (
       <TextField {...textFieldProps} />
@@ -148,5 +148,4 @@ export default class Autosuggest extends Component {
       </div>
     );
   }
-
 }
