@@ -4,14 +4,14 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
 import classnames from 'classnames';
+import capitalize from 'lodash/capitalize';
 
 export default class Button extends Component {
-
   static displayName = 'Button';
 
   static propTypes = {
     color: PropTypes.oneOf([
-      'pink', 'blue', 'gray', 'transparent'
+      'pink', 'blue', 'gray', 'transparent', 'white'
     ]).isRequired,
     children: PropTypes.oneOfType([
       PropTypes.arrayOf(PropTypes.node),
@@ -22,40 +22,34 @@ export default class Button extends Component {
       PropTypes.func,
       PropTypes.string
     ]),
+    ghost: PropTypes.bool,
     loading: PropTypes.bool,
     fullWidth: PropTypes.bool
   };
 
   static defaultProps = {
     className: '',
+    ghost: false,
     loading: false,
     fullWidth: false,
     component: 'button'
   };
 
-  constructor() {
-    super();
-
-    this.storeButtonReference = this.storeButtonReference.bind(this);
-  }
-
-  storeButtonReference(button) {
+  storeButtonReference = button => {
     if (button !== null) {
       this.button = button;
     }
   }
 
   render() {
-    const { color, className, loading, fullWidth, children, component, ...restProps } = this.props;
+    const { color, ghost, className, loading, fullWidth, children, component, ...restProps } = this.props;
 
     const combinedProps = {
       className: classnames(styles.root, className, {
         [styles.loading]: loading,
         [styles.fullWidth]: fullWidth,
-        [styles.root_isPink]: color === 'pink',
-        [styles.root_isBlue]: color === 'blue',
-        [styles.root_isGray]: color === 'gray',
-        [styles.root_isTransparent]: color === 'transparent'
+        [styles.root_isGhost]: ghost && color !== 'transparent',
+        [styles[`root_is${capitalize(color)}`]]: color
       }),
       disabled: loading,
       ref: this.storeButtonReference,
@@ -64,5 +58,4 @@ export default class Button extends Component {
 
     return React.createElement(component, combinedProps, children);
   }
-
 }
