@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-
 import Checkbox from './Checkbox';
+import * as sketch from './Checkbox.sketch';
 import demoStyles from './Checkbox.demo.less';
 
 class CheckboxContainer extends Component {
@@ -47,6 +47,7 @@ export default {
   title: 'Checkbox',
   component: Checkbox,
   container: CheckboxContainer,
+  sketch,
   initialProps: {
     id: 'stillInRole',
     label: 'Still in role',
