@@ -30,7 +30,7 @@ describe('Dropdown', () => {
   beforeEach(() => {
     errors = [];
 
-    sinon.stub(console, 'error', errorMessage => {
+    sinon.stub(console, 'error').callsFake(errorMessage => {
       errors.push(errorMessage);
     });
   });
