@@ -1,15 +1,12 @@
-import styles from './UserAccount.less';
-
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import ScrollLock from 'react-scrolllock';
-
 import ChevronIcon from '../../ChevronIcon/ChevronIcon';
 import ScreenReaderOnly from '../../ScreenReaderOnly/ScreenReaderOnly';
 import UserAccountMenu from '../UserAccountMenu/UserAccountMenu';
 import { AUTHENTICATED, UNAUTHENTICATED, AUTH_PENDING } from '../../private/authStatusTypes';
-
 import smallDeviceOnly from '../../private/smallDeviceOnly';
+import styles from './UserAccount.less';
 
 const calculateMobileMenuLabel = (authenticationStatus, userName) => {
   if (authenticationStatus === AUTH_PENDING) {
@@ -34,8 +31,13 @@ export default class UserAccount extends Component {
     userName: PropTypes.string,
     linkRenderer: PropTypes.func.isRequired,
     returnUrl: PropTypes.string,
-    activeTab: PropTypes.string
+    activeTab: PropTypes.string,
+    onMenuToggle: PropTypes.func
   };
+
+  static defaultProps = {
+    onMenuToggle: () => {}
+  }
 
   constructor(props) {
     super(props);
@@ -46,7 +48,10 @@ export default class UserAccount extends Component {
   }
 
   handleMenuToggleClick = () => {
-    this.setState(state => ({ menuOpen: !state.menuOpen }));
+    this.setState(state => {
+      this.props.onMenuToggle({ open: !state.menuOpen });
+      return { menuOpen: !state.menuOpen };
+    });
   };
 
   handleMenuClick = () => {
