@@ -4,23 +4,9 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
 import classnames from 'classnames';
-import range from 'lodash/range';
 
+import getYearOptions from './getYearOptions';
 import Dropdown from '../../Dropdown/Dropdown';
-
-const getYearOptions = () => {
-  const maxYear = new Date().getFullYear();
-  const minYear = maxYear - 100;
-
-  return range(maxYear, minYear - 1).map(value => {
-    const stringValue = String(value);
-
-    return {
-      value: stringValue,
-      label: stringValue
-    };
-  });
-};
 
 const months = [
   { value: '1', label: 'Jan' },
@@ -36,10 +22,8 @@ const months = [
   { value: '11', label: 'Nov' },
   { value: '12', label: 'Dec' }
 ];
-const years = getYearOptions();
 
 export default class CustomMonthPicker extends Component {
-
   static displayName = 'CustomMonthPicker';
 
   static propTypes = {
@@ -51,7 +35,10 @@ export default class CustomMonthPicker extends Component {
     }),
     valid: PropTypes.bool,
     className: PropTypes.string,
-    id: PropTypes.string
+    id: PropTypes.string,
+    minYear: PropTypes.number.isRequired,
+    maxYear: PropTypes.number.isRequired,
+    ascendingYears: PropTypes.bool.isRequired
   };
 
   static defaultProps = {
@@ -59,7 +46,7 @@ export default class CustomMonthPicker extends Component {
     className: ''
   };
 
-  constructor() {
+  constructor({ minYear, maxYear, ascendingYears }) {
     super();
 
     this.handleMonthChange = this.handleMonthChange.bind(this);
@@ -68,6 +55,22 @@ export default class CustomMonthPicker extends Component {
     this.storeYearReference = this.storeYearReference.bind(this);
     this.handleBlur = this.handleBlur.bind(this);
     this.blurIfNotFocussed = this.blurIfNotFocussed.bind(this);
+
+    this.yearOptions = getYearOptions(minYear, maxYear, ascendingYears);
+  }
+
+  componentWillUpdate(newProps) {
+    if (
+      ['minYear', 'maxYear', 'ascendingYears'].filter(
+        key => newProps[key] !== this.props[key]
+      )
+    ) {
+      this.yearOptions = getYearOptions(
+        newProps.minYear,
+        newProps.maxYear,
+        newProps.ascendingYears
+      );
+    }
   }
 
   storeMonthReference(input) {
@@ -157,7 +160,7 @@ export default class CustomMonthPicker extends Component {
           }}
         />
         <Dropdown
-          options={years}
+          options={this.yearOptions}
           className={styles.dropdown}
           valid={valid}
           message={false}
