@@ -16,7 +16,7 @@ describe('MonthPicker', () => {
   beforeEach(() => {
     errors = [];
 
-    sinon.stub(console, 'error', errorMessage => {
+    sinon.stub(console, 'error').callsFake(errorMessage => {
       errors.push(errorMessage);
     });
   });
