@@ -21,7 +21,7 @@ describe('NativeMonthPicker', () => {
   beforeEach(() => {
     errors = [];
 
-    sinon.stub(console, 'error', errorMessage => {
+    sinon.stub(console, 'error').callsFake(errorMessage => {
       errors.push(errorMessage);
     });
   });
