@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-
 import TextField from './TextField';
 import styles from './TextField.less';
+import * as sketch from './TextField.sketch';
 import classnames from 'classnames';
 import fieldMessageOptions from '../private/FieldMessage/FieldMessage.demo';
 import fieldLabelOptions from '../private/FieldLabel/FieldLabel.demo';
@@ -57,6 +57,7 @@ export default {
   title: 'Text Field',
   component: TextField,
   container: TextFieldContainer,
+  sketch,
   initialProps: {
     id: 'firstName',
     label: 'First Name',
